+++ networks/archive/bc.py
@@ -1,4 +1,4 @@
-from agents.base_network import Network
+from networks.base_network import Network
 from helpers.environment import ObservationSpace, ActionSpace
 
 import os
