+++ networks/termination_critic.py
@@ -1,5 +1,5 @@
 from helpers.environment import ObservationSpace, ActionSpace
-from agents.base_network import Network
+from networks.base_network import Network
 
 import os
 import time
